% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95 Percent Confidence Interval}
\usage{
myci(x)
}
\arguments{
\item{x}{Distribution}
}
\value{
Confidence interval for the mean of the distribution
}
\description{
95 Percent Confidence Interval
}
\examples{
\dontrun{x <- rnorm(20, 10, 10); myci(x=x)}
}
