% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Title}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Iterations}

\item{lambda}{Lambda in Poisson Distribution}

\item{...}{}
}
\value{

}
\description{
Title
}
