% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max Likelihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Logarithmic function input}

\item{x}{Set of data}

\item{param}{Sequence of values of which to range the plot}

\item{...}{Other graphical properties and parameters for the plot output}
}
\value{
Returns a plot representing the logarithmic function alongside a point where the maximum likelihood is
}
\description{
Max Likelihood function
}
\examples{
\dontrun{logbin=function(x,param) log(dbinom(x,prob=param,size=20));
logpoiss=function(x,param) log(dpois(x,lambda=param));
logexp=function(x,param) log(dexp(x,rate=param));
mymaxlik(x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000),lfun=logbin,main="Binomial",cex.main=1)}
}
