% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Function for printing probability on curves}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Standard Deviation}

\item{a}{}
}
\value{
Displays a curve with a shaded area between the curve and the x axis
from -inf to x = a, and prints the probability P(X<=a)
}
\description{
Function for printing probability on curves
}
