% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{P-value Areas}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{An initial T value}

\item{xmax}{Maximum x on graph}

\item{n}{Sample size}

\item{alpha}{Confidence Alpha}
}
\value{
Returns a plot distributed normally with quantiles
marked alongside a shaded region of probability estimates
}
\description{
P-value Areas
}
\examples{
\dontrun{tcalc=(mean(x1)-24)/(sd(x1)/sqrt(30));mypvalue(tcalc,n=30,alpha=0.05)}
}
