% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{A function for showing a barplot of samples}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Number of iterations}

\item{time}{Delay between iterations}
}
\value{
Set of iter barplots
}
\description{
A function for showing a barplot of samples
}
