% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadPlot.R
\name{quadPlot}
\alias{quadPlot}
\title{A function for plotting a quadratic}
\usage{
quadPlot(x, qlm)
}
\arguments{
\item{x}{The diameter}
}
\value{
Response i.e. estimated Height
}
\description{
A function for plotting a quadratic
}
\examples{
\dontrun{quad.lm<-lm(y~x,data=df); myplot(x=15, qlm=quad.lm)}
}
