% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Iterations}

\item{lambda}{Lambda in Poisson Distribution}

\item{...}{Further graphical parameters}
}
\value{
Returns a histogram of the sample mean, a barplot of a sampled y,
and a probability function for poisson distribution
}
\description{
CLT Poisson
}
