% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max Likelihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Logarithmic function input}

\item{x}{Set of data}

\item{param}{Sequence of values of which to range the plot}

\item{...}{Other graphical properties and parameters for the plot output}
}
\value{
Returns a list
}
\description{
Max Likelihood function
}
